refactor: 클래스 이름 변경, 필요없는 depth 제거
