+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -14,11 +14,10 @@ class ShoppingApplication : Application() {
         super.onCreate()
         container = DefaultAppContainer()
 
-        container.injectedComponentContainer.add(
+        container.add(
             InjectedComponent.InjectedSingletonComponent(ProductRepository::class, DefaultProductRepository()),
         )
-
-        container.injectedComponentContainer.add(
+        container.add(
             InjectedComponent.InjectedSingletonComponent(CartRepository::class, DefaultCartRepository()),
         )
     }
